Adjust header to design - Categories in categoriesList
