+++ src/Pages/Category.css
@@ -34,7 +34,7 @@ a {
   margin-top: 3%;
   margin-bottom: 3%;
   display: grid;
-  grid-template-columns: 1fr;
+  grid-template-columns: 1fr 1fr;
   gap: 0;
   text-align: left;
 }
@@ -79,22 +79,18 @@ a {
 .photoInformation {
   display: flex;
   flex-direction: column;
-  padding: 12px;
   flex-basis: 50%;
   gap: 0.5em;
 }
 
 .icon-wrapper {
-  /*
-  position: absolute;
-  right: 1rem;
-  top: 0.5rem;
-  */
   border-radius: 50%;
   border: 1px solid #fff;
   padding: 0.2rem;
   display: flex;
   align-self: flex-end;
+  font-size: 15px;
+  margin:3px;
 }
 
 .photoStatus {
@@ -127,8 +123,9 @@ a {
   height: 30px;
   margin-top: 3%;
   margin-left: 3%;
-  border: 1px solid black;
+  border: 1px solid #b50be5;
   display: flex;
+  background-color:  rgb(177, 57, 103);
 }
 
 #searchInput {
@@ -140,6 +137,7 @@ a {
   background: transparent;
   padding: 10px;
   padding-bottom: 12px;
+  color: white;
 }
 
 #searchBtn {
