+++ src/Pages/Home.js
@@ -2,33 +2,12 @@ import { NavLink } from 'react-router-dom';
 import { VscFoldDown } from 'react-icons/vsc';
 import { useDispatch } from 'react-redux';
 import { FaArrowRight } from 'react-icons/fa';
+import categories from './categoriesList';
 import { filterCategory } from '../Redux/PhotosSlice';
 import './Home.css';
 
 const Home = () => {
   const dispatch = useDispatch();
-
-  const categories = [
-    { categoryCode: 'backgrounds', categoryName: 'backgrounds', page: 0 },
-    { categoryCode: 'fashion', categoryName: 'fashion', page: 0 },
-    { categoryCode: 'nature', categoryName: 'nature', page: 0 },
-    { categoryCode: 'science', categoryName: 'science', page: 0 },
-    { categoryCode: 'education', categoryName: 'education', page: 0 },
-    { categoryCode: 'feelings', categoryName: 'feelings', page: 0 },
-    { categoryCode: 'health', categoryName: 'health', page: 0 },
-    { categoryCode: 'people', categoryName: 'people', page: 0 },
-    { categoryCode: 'religion', categoryName: 'religion', page: 0 },
-    { categoryCode: 'places', categoryName: 'places', page: 0 },
-    { categoryCode: 'animals', categoryName: 'animals', page: 0 },
-    { categoryCode: 'industry', categoryName: 'industry', page: 0 },
-    { categoryCode: 'computer', categoryName: 'computer', page: 0 },
-    { categoryCode: 'food', categoryName: 'food', page: 0 },
-    { categoryCode: 'sports', categoryName: 'sports', page: 0 },
-    { categoryCode: 'transportation', categoryName: 'transportation', page: 0 },
-    { categoryCode: 'travel', categoryName: 'travel', page: 0 },
-    { categoryCode: 'buildings', categoryName: 'buildings', page: 0 },
-    { categoryCode: 'music', categoryName: 'music', page: 0 },
-  ];
 
   const handleClick = (code, page, name) => {
     const parameters = {
@@ -42,10 +21,12 @@ const Home = () => {
   return (
     <div id="homeContainer">
       <header>
-        <div id="headerImage" />
+        <div id="headerImage" className="filtered">
+          <img src="../Resources/Images/header2.jpg" alt="header" />
+        </div>
         <div id="headerText">
-          <h1>WELLCOME TO MiEVENT</h1>
-          <p id="headerDescription">A place where you can find your dream event, anywhere</p>
+          <h1>WELCOME TO PHOTO STORE</h1>
+          <p id="headerDescription">A place where you can find the best photos for free</p>
           <a
             id="discoverAction"
             href="#categoryGrid"
@@ -54,9 +35,7 @@ const Home = () => {
           </a>
         </div>
       </header>
-      <h2 id="gridTitle">EVENTS BY COUNTRY</h2>
-      <h2 id="desktopGridTitle">CHOOSE YOUR COUNTRY!</h2>
-      <hr />
+      <h2 id="gridTitle">PHOTOS BY CATEGORY</h2>
       <div id="categoryGrid">
         {categories.map((category) => (
           <div
