+++ src/Pages/categoriesList.js
@@ -1,0 +1,23 @@
+const categories = [
+  { categoryCode: 'backgrounds', categoryName: 'backgrounds', page: 0 },
+  { categoryCode: 'fashion', categoryName: 'fashion', page: 0 },
+  { categoryCode: 'nature', categoryName: 'nature', page: 0 },
+  { categoryCode: 'science', categoryName: 'science', page: 0 },
+  { categoryCode: 'education', categoryName: 'education', page: 0 },
+  { categoryCode: 'feelings', categoryName: 'feelings', page: 0 },
+  { categoryCode: 'health', categoryName: 'health', page: 0 },
+  { categoryCode: 'people', categoryName: 'people', page: 0 },
+  { categoryCode: 'religion', categoryName: 'religion', page: 0 },
+  { categoryCode: 'places', categoryName: 'places', page: 0 },
+  { categoryCode: 'animals', categoryName: 'animals', page: 0 },
+  { categoryCode: 'industry', categoryName: 'industry', page: 0 },
+  { categoryCode: 'computer', categoryName: 'computer', page: 0 },
+  { categoryCode: 'food', categoryName: 'food', page: 0 },
+  { categoryCode: 'sports', categoryName: 'sports', page: 0 },
+  { categoryCode: 'transportation', categoryName: 'transportation', page: 0 },
+  { categoryCode: 'travel', categoryName: 'travel', page: 0 },
+  { categoryCode: 'buildings', categoryName: 'buildings', page: 0 },
+  { categoryCode: 'music', categoryName: 'music', page: 0 },
+];
+
+export default categories;
